refactor(routes): 保護ルートを(app)配下に集約し認証/リダイレクトを整理
- 認証必須領域の親ルートとして `/(app)` を追加し、未認証時は `/login` へ `redirect=location.href` 付きでリダイレクト
- `__root` から認証ガードを除去して、ガード責務を `/(app)` 側に集約
- `/login` の `redirect` を正規化（`javascript:`/`data:` や別オリジン等を拒否）し、既ログイン時は `redirect({ href })`、ログイン成功後は `router.history.push()` で遷移
- demo/ui/settings などのルートを `/(app)` グループ配下へ移動し、生成ルートツリーも追従更新
