+++ src/routes/(app)/demo.chat-t.tsx
@@ -1,0 +1,15 @@
+import { createFileRoute } from '@tanstack/react-router'
+import { Chat } from '@/features/chat/components/chat'
+
+export const Route = createFileRoute('/(app)/demo/chat-t')({
+  component: RouteComponent,
+  loader: () => ({ crumb: 'AIチャット(TanStack AI)' }),
+})
+
+function RouteComponent() {
+  return (
+    <div>
+      <Chat />
+    </div>
+  )
+}
