+++ src/routes/(app)/demo.route.tsx
@@ -1,0 +1,10 @@
+import { Outlet, createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/(app)/demo')({
+  component: RouteComponent,
+  loader: () => ({ crumb: 'Demo' }),
+})
+
+function RouteComponent() {
+  return <Outlet />
+}
