+++ src/routes/(app)/ui/route.tsx
@@ -1,0 +1,10 @@
+import { Outlet, createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/(app)/ui')({
+  component: RouteComponent,
+  loader: () => ({ crumb: 'UI' }),
+})
+
+function RouteComponent() {
+  return <Outlet />
+}
