+++ src/routes/__root.tsx
@@ -1,8 +1,4 @@
-import {
-  Outlet,
-  createRootRouteWithContext,
-  redirect,
-} from '@tanstack/react-router'
+import { Outlet, createRootRouteWithContext } from '@tanstack/react-router'
 import { TanStackRouterDevtoolsPanel } from '@tanstack/react-router-devtools'
 import { TanStackDevtools } from '@tanstack/react-devtools'
 import { aiDevtoolsPlugin } from '@tanstack/react-ai-devtools'
@@ -23,19 +19,6 @@ interface MyRouterContext {
 
 export const Route = createRootRouteWithContext<MyRouterContext>()({
   component: Component,
-  beforeLoad: ({ context, location: { pathname, href } }) => {
-    if (pathname === '/') {
-      // トップページは常にアクセス可能
-    } else if (pathname === '/login') {
-      // Login ページは常にアクセス可能
-    } else if (!context.auth.isAuthenticated) {
-      // 認証されていないユーザーが保護されたページにアクセスしようとした場合、
-      // /login にリダイレクト
-      throw redirect({ to: '/login', search: { redirect: href } })
-    } else {
-      // 認証されているユーザーはアクセス可能
-    }
-  },
 })
 
 function Component() {
