--- src/routes/demo.route.tsx
@@ -1,10 +1,0 @@
-import { Outlet, createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/demo')({
-  component: RouteComponent,
-  loader: () => ({ crumb: 'Demo' }),
-})
-
-function RouteComponent() {
-  return <Outlet />
-}
